requirements.txt
feedparser
requests
openai
newspaper3k
